requirements.txt
black>=23.3.0
flake8>=3.7.8
mkdocs>=1.4.2
mkdocs-material>=9.1.6
python-slugify==4.0.1
GitPython>=3.1.31
PyYaml>=6.0
